seperate builtin_func
