+++ builtin_func/export_unset.c
@@ -65,27 +65,3 @@ t_bool	ft_unset(t_node *node)
 	}
 	return (FALSE);
 }
-
-// 노드 삭제 함수
-void	delete_node(t_list **head, t_list *node_to_delete)
-{
-	t_list	*current;
-	t_list	*prev;
-
-	current = *head;
-	prev = NULL;
-	while (current != NULL)
-	{
-		if (current == node_to_delete)
-		{
-			if (prev == NULL)
-				*head = current->next;
-			else
-				prev->next = current->next;
-			free(current);
-			break ;
-		}
-		prev = current;
-		current = current->next;
-	}
-}
