+++ unit_test/test.c
@@ -52,6 +52,7 @@ void	init_envp(char **envp)
 		ft_lstadd_back(env, ft_lstnew(ft_strdup(envp[idx])));
 		idx++;
 	}
+	ft_lstadd_back(env, ft_lstnew(NULL));
 }
 
 t_list	**get_envp(void)
@@ -69,12 +70,15 @@ void	set_envp(char *pos, char *pwd)
 
 	pos_len = ft_strlen(pos);
 	temp = getenv_list(pos, pos_len, get_envp());
+	// 없을 경우 추가
 	if (!temp)
 	{
 		newpwd = ft_strnjoin(pos, pwd);
-		ft_lstadd_back(get_envp(), ft_lstnew(ft_strdup(newpwd)));
+		// newpwd = OLDPWD=/qwdgsjghskjdhg/aeoghaeghaeghigauhuih
+		ft_lstadd_back(get_envp(), ft_lstnew(newpwd));
 		free(newpwd);
 	}
+	// 있을 경우 삭제, 프리 하고 추가
 	else
 	{
 		newpwd = ft_strnjoin(pos, pwd);
@@ -114,7 +118,7 @@ void	ft_env(void)
 	idx = 0;
 	env = get_envp();
 	temp = *env;
-	while (temp != 0)
+	while (temp->next != 0)
 	{
 		printf("%s\n", temp->content);
 		temp = temp->next;
@@ -125,21 +129,20 @@ int main(int argc, char **argv, char **envp)
 {
 	(void)argc;
 	(void)argv;
-	
 	char *test = "test99";
 	char	*newpwd;
 	char	path[PATH_MAX];
 
 	init_envp(envp);
-	ft_env();
+	// ft_env();
 	newpwd = getcwd(path, PATH_MAX);
 	if (chdir(test) != 0)
 	{
 		printf("error");
 		exit(1);
 	}
-	set_envp("OLDPWD", newpwd);
+	set_envp("OLDPWD", newpwd); // 새로 추가되는 파트
 	newpwd = getcwd(path, PATH_MAX);
-	set_envp("PWD", newpwd);
-	ft_env();
+	set_envp("PWD", newpwd); // 수정파트 -> 동작O
+	ft_env(); // -> PWD가 갱신된것을 확인 / OLDPWD가 없음 -> 추가 -> (NULL)
 }
